rodski-web/requirements.txt
flask>=2.3.0
pyyaml>=6.0
